% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook_functions.R
\name{check_ID_data_level}
\alias{check_ID_data_level}
\title{Check if the data is at the ID level}
\usage{
check_ID_data_level(data)
}
\arguments{
\item{data}{A character string specifying the name of the dataset.}
}
\value{
An integer indicating the outcome:
\itemize{
\item 0 - There is no ID variable that is Tricot-Defined in this data.
\item 1 - No key columns are defined in the dataset.
\item 2 - Only ID level data can be used for this data. This is data where there is a unique row for each ID variable given.
\item 3 - Success. The dataset is Tricot-defined and at the ID level.
}
Additionally, a message is printed describing the result.
}
\description{
This function checks whether the specified dataset is Tricot-defined
and whether it is at the ID level. A dataset is considered
"at the ID level" if one of its key columns contains only the ID identifier.
}
