% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook_functions.R
\name{check_variety_data_level}
\alias{check_variety_data_level}
\title{Check if the data is at the variety level}
\usage{
check_variety_data_level(data)
}
\arguments{
\item{data}{A character string specifying the name of the dataset.}
}
\value{
An integer indicating the outcome:
\itemize{
\item 0 - There is no variety variable that is Tricot-Defined in this data.
\item 1 - No key columns are defined in the dataset.
\item 2 - "Only variety level data can be used for this data. This is data where there is a unique row for each variety given.
\item 3 - Success. The dataset is Tricot-defined and at the variety level.
}
Additionally, a message is printed describing the result.
}
\description{
This function checks whether the specified dataset is Tricot-defined
and whether it is at the variety level. A dataset is considered
"at the variety level" if one of its key columns contains only the variety identifier.
}
