% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{create_season_length_definitions}
\alias{create_season_length_definitions}
\title{Get season length definitions (internal)}
\usage{
create_season_length_definitions(length = NULL)
}
\arguments{
\item{length}{A definition node.}
}
\value{
A list with \code{seasonal_length$end_type}.
}
\description{
Get season length definitions (internal)
}
\keyword{internal}
