% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databook_functions.R
\name{create_tricot_data}
\alias{create_tricot_data}
\title{Create and Structure Tricot Data at Multiple Levels}
\usage{
create_tricot_data(
  output_data_levels,
  id_level_data = "",
  id_col = "id",
  good_suffixes = c("_pos", "_best"),
  bad_suffixes = c("_neg", "_worst"),
  na_candidates = c("Not observed", "Not scored", NA_character_)
)
}
\arguments{
\item{output_data_levels}{A tibble from \code{instatExtras}' \code{summarise_data_levels()} that describes the structure of input data sets.
Must contain at least one dataset at the "id" level.}

\item{id_level_data}{Optional string. If no dataset is automatically detected at the "id" level, this can be used
to specify the dataset name manually.}

\item{id_col}{The name of the ID column in the dataset (default is \code{"id"}).}

\item{good_suffixes}{Character vector of suffixes used for positive trait rankings (e.g., \code{"_pos"}, \code{"_best"}).}

\item{bad_suffixes}{Character vector of suffixes used for negative trait rankings (e.g., \code{"_neg"}, \code{"_worst"}).}

\item{na_candidates}{Character vector of values considered as missing (e.g., \code{"Not observed"}). Must be a single value if supplied.}
}
\value{
The function returns nothing explicitly. It modifies the \code{data_book} object by importing, transforming,
linking, and defining structured tricot data and storing relevant ranking objects.
}
\description{
This function prepares and structures tricot data by detecting the appropriate data level
(ID, ID-variety, ID-variety-trait, or variety) and creating the necessary data frames and metadata.
It also defines the data within a \code{data_book} object for further analysis, and generates ranking objects.
}
\details{
This function does the following:
\enumerate{
\item Detects and defines the ID-level data using the \code{define_as_tricot()} function.
\item If no ID-variety data is available, it creates it using \code{instatExtras::pivot_tricot()} based on detected structure.
\item Defines ID-variety and variety-level datasets in the \code{data_book} with appropriate metadata.
\item Adds a link between the ID-variety and variety datasets.
\item Generates \code{rankings_list} and \code{grouped_rankings_list} objects using \code{gosset::rank_numeric()} and stores them in the \code{data_book}.
}
}
