% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_book.R
\docType{class}
\name{DataBook}
\alias{DataBook}
\title{DataBook Class}
\format{
An R6 class object.
}
\usage{
DataBook$new(data_tables = list(), instat_obj_metadata = list(), 
                    data_tables_variables_metadata = rep(list(data.frame()), length(data_tables)),
                    data_tables_metadata = rep(list(list()), length(data_tables)),
                    data_tables_filters = rep(list(list()), length(data_tables)),
                    data_tables_column_selections = rep(list(list()), length(data_tables)),
                    imported_from = as.list(rep("", length(data_tables))),
                    messages = TRUE, convert = TRUE, create = TRUE)
}
\description{
DataBook Class

DataBook Class
}
\details{
An R6 class to manage a collection of data tables along with their metadata and other associated properties.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data_objects_changed}}{Logical indicating whether the data objects have changed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataBook-new}{\code{DataBook$new()}}
\item \href{#method-DataBook-standardise_country_names}{\code{DataBook$standardise_country_names()}}
\item \href{#method-DataBook-define_as_climatic}{\code{DataBook$define_as_climatic()}}
\item \href{#method-DataBook-define_corruption_outputs}{\code{DataBook$define_corruption_outputs()}}
\item \href{#method-DataBook-define_red_flags}{\code{DataBook$define_red_flags()}}
\item \href{#method-DataBook-define_as_procurement}{\code{DataBook$define_as_procurement()}}
\item \href{#method-DataBook-define_as_procurement_country_level_data}{\code{DataBook$define_as_procurement_country_level_data()}}
\item \href{#method-DataBook-get_CRI_component_column_names}{\code{DataBook$get_CRI_component_column_names()}}
\item \href{#method-DataBook-get_red_flag_column_names}{\code{DataBook$get_red_flag_column_names()}}
\item \href{#method-DataBook-get_CRI_column_names}{\code{DataBook$get_CRI_column_names()}}
\item \href{#method-DataBook-get_corruption_column_name}{\code{DataBook$get_corruption_column_name()}}
\item \href{#method-DataBook-clone}{\code{DataBook$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataBook object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$new(
  data_tables = list(),
  instat_obj_metadata = list(),
  data_tables_variables_metadata = rep(list(data.frame()), length(data_tables)),
  data_tables_metadata = rep(list(list()), length(data_tables)),
  data_tables_filters = rep(list(list()), length(data_tables)),
  data_tables_column_selections = rep(list(list()), length(data_tables)),
  imported_from = as.list(rep("", length(data_tables))),
  messages = TRUE,
  convert = TRUE,
  create = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_tables}}{A list of data frames to be included in the DataBook.}

\item{\code{instat_obj_metadata}}{Metadata for the instat object.}

\item{\code{data_tables_variables_metadata}}{A list of data frames, each containing metadata for the corresponding data table.}

\item{\code{data_tables_metadata}}{A list of lists, each containing metadata for the corresponding data table.}

\item{\code{data_tables_filters}}{A list of lists, each containing filter information for the corresponding data table.}

\item{\code{data_tables_column_selections}}{A list of lists, each containing column selection information for the corresponding data table.}

\item{\code{imported_from}}{A list of strings indicating the source from which each data table was imported.}

\item{\code{messages}}{A boolean indicating whether to display messages.}

\item{\code{convert}}{A boolean indicating whether to perform data conversion.}

\item{\code{create}}{A boolean indicating whether to create new data objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-standardise_country_names"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-standardise_country_names}{}}}
\subsection{Method \code{standardise_country_names()}}{
Standardize country names in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$standardise_country_names(data_name, country_columns = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{country_columns}}{A vector of column names containing country data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-define_as_climatic"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-define_as_climatic}{}}}
\subsection{Method \code{define_as_climatic()}}{
Define a data table as climatic data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$define_as_climatic(data_name, types, key_col_names, key_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{types}}{A vector specifying the types of climatic data.}

\item{\code{key_col_names}}{A vector of column names to be used as keys.}

\item{\code{key_name}}{The name of the key.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-define_corruption_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-define_corruption_outputs}{}}}
\subsection{Method \code{define_corruption_outputs()}}{
Define corruption output columns in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$define_corruption_outputs(data_name, output_columns = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{output_columns}}{A vector of column names to be defined as corruption outputs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-define_red_flags"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-define_red_flags}{}}}
\subsection{Method \code{define_red_flags()}}{
Define red flag columns in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$define_red_flags(data_name, red_flags = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{red_flags}}{A vector of column names to be defined as red flags.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-define_as_procurement"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-define_as_procurement}{}}}
\subsection{Method \code{define_as_procurement()}}{
Define a data table as procurement data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$define_as_procurement(
  data_name,
  primary_types = c(),
  calculated_types = c(),
  country_data_name,
  country_types,
  auto_generate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{primary_types}}{A vector of primary types of procurement data.}

\item{\code{calculated_types}}{A vector of calculated types of procurement data.}

\item{\code{country_data_name}}{The name of the country-level data table.}

\item{\code{country_types}}{A vector of types for the country-level data.}

\item{\code{auto_generate}}{A boolean indicating whether to auto-generate procurement types.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-define_as_procurement_country_level_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-define_as_procurement_country_level_data}{}}}
\subsection{Method \code{define_as_procurement_country_level_data()}}{
Define a data table as procurement country-level data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$define_as_procurement_country_level_data(
  data_name,
  contract_level_data_name,
  types = c(),
  auto_generate = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{contract_level_data_name}}{The name of the contract-level data table.}

\item{\code{types}}{A vector of types for the procurement country-level data.}

\item{\code{auto_generate}}{A boolean indicating whether to auto-generate procurement types.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-get_CRI_component_column_names"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-get_CRI_component_column_names}{}}}
\subsection{Method \code{get_CRI_component_column_names()}}{
Get the names of CRI component columns in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$get_CRI_component_column_names(data_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of CRI component column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-get_red_flag_column_names"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-get_red_flag_column_names}{}}}
\subsection{Method \code{get_red_flag_column_names()}}{
Get the names of red flag columns in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$get_red_flag_column_names(data_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of red flag column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-get_CRI_column_names"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-get_CRI_column_names}{}}}
\subsection{Method \code{get_CRI_column_names()}}{
Get the names of CRI columns in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$get_CRI_column_names(data_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of CRI column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-get_corruption_column_name"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-get_corruption_column_name}{}}}
\subsection{Method \code{get_corruption_column_name()}}{
Get the name of the corruption column in the specified data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$get_corruption_column_name(data_name, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data table.}

\item{\code{type}}{The type of the corruption column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The name of the corruption column.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBook-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataBook-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBook$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
