% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayDaily.R
\name{DisplayDaily}
\alias{DisplayDaily}
\title{Display Daily Meteorological Data}
\usage{
DisplayDaily(
  Datain,
  Stations,
  Variables,
  option = 1,
  Years,
  Misscode,
  Tracecode,
  Zerocode,
  Fileout = NA,
  monstats = c("min", "mean", "median", "max", "IQR", "sum")
)
}
\arguments{
\item{Datain}{A data frame containing meteorological data, with one column for each variable, one row for each date, and the following columns (case-sensitive):
\itemize{
\item \code{Station}: names or identifiers for stations
\item EITHER: a \code{Date} column (yyyy-mm-dd format, \code{option = 1})
\item OR: separate \code{Year}, \code{Month}, and \code{Day} columns (\code{option = 2})
}}

\item{Stations}{A character vector of station names to view (e.g., \code{"KUND0002"}, \code{c("Paris", "London")}). If missing, all stations in \code{Datain} are selected.}

\item{Variables}{A character vector of column names for the variables to display (e.g., \code{"Rain"}, \code{"TMin"}, \code{"RH"}). If missing, all variables except metadata columns are selected.}

\item{option}{An integer (1 or 2) indicating date format:
\itemize{
\item \code{1}: \code{Date} column format (yyyy-mm-dd)
\item \code{2}: separate \code{Year}, \code{Month}, and \code{Day} columns.
}}

\item{Years}{An integer vector of years to view (default is all years in \code{Datain}).}

\item{Misscode}{Character string representing the display value for missing data (e.g., "-").}

\item{Tracecode}{Character string for trace amounts of rainfall (default "tr").}

\item{Zerocode}{Character string for zero values (default "--").}

\item{Fileout}{Optional file path for saving the output table. If \code{NA}, results are printed to the console.}

\item{monstats}{A character vector of monthly statistics to compute, options include \code{"min"}, \code{"mean"}, \code{"median"}, \code{"max"}, \code{"IQR"}, and \code{"sum"}.}
}
\value{
This function prints a formatted daily table for each specified station, variable, and year. If a file path is specified in \code{Fileout}, the output is saved to that file.
}
\description{
Display daily meteorological data in a readable table format.
This function was developed by Helen Greatrex as part of the SSC-RCLIM package. It formats daily meteorological data for easy viewing, displaying values for each day of the month and summarising monthly statistics. The function is similar in display format to the INSTAT "DisplayDaily" feature.
}
\details{
This function checks for the presence of required columns and renames variables as needed to ensure compatibility with the \code{Datain} format. Missing data, trace rainfall, and zero values are replaced with user-defined codes. Monthly statistics are computed for each month. The function fills in any missing dates within the specified range for seamless daily reporting.
}
