\name{dodoma}
\docType{data}
\alias{dodoma}

\title{Dodoma example dataset}

\description{A dataset used in package tests representing weather/plot observations for the Dodoma case study.}

\usage{data(dodoma, package = "databook")}

\format{A data frame with multiple columns used by tests.}

\source{Provided with the package for testing purposes.}
