% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{EDI}
\alias{EDI}
\title{Calculate Extremal Dependency Index}
\usage{
EDI(x, y, frcst.type, obs.type, ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Predicted values.}

\item{frcst.type}{Character. The type of forecast (e.g., "categorical").}

\item{obs.type}{Character. The type of observation (e.g., "categorical").}

\item{...}{Additional arguments passed to \code{verification::verify}.}
}
\value{
The extremal dependency index.
}
\description{
Computes the extremal dependency index (EDI) using the \code{verification::verify} function.
}
