% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{EDS}
\alias{EDS}
\title{Calculate Extreme Dependency Score}
\usage{
EDS(x, y, frcst.type, obs.type, ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Predicted values.}

\item{frcst.type}{Character. The type of forecast (e.g., "categorical").}

\item{obs.type}{Character. The type of observation (e.g., "categorical").}

\item{...}{Additional arguments passed to \code{verification::verify}.}
}
\value{
The extreme dependency score or NA if unsupported.
}
\description{
Computes the extreme dependency score (EDS) using the \code{verification::verify} function.
Returns NA with a warning if unsupported by the verification package.
}
