% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{get_definition_column}
\alias{get_definition_column}
\title{Get a column from definitions_year}
\usage{
get_definition_column(definitions_year, column_name)
}
\arguments{
\item{definitions_year}{name of the definitions found with
\code{get_r_instat_definitions(calculations_data)}}

\item{column_name}{A character string giving the name of the column
to retrieve. If \code{NULL}, the function returns \code{NULL}.}
}
\value{
A vector containing the requested column from
\code{definitions_year}, or \code{NULL} if \code{column_name} is
\code{NULL}.
}
\description{
Retrieves a column from \code{definitions_year}. If multiple columns
exist with the same name, the most recently created column (i.e. the
last matching column) is returned.
}
\details{
This function is intended for use in situations where duplicate column
names may occur in \code{definitions_year}. In such cases, the function
assumes that the most recently created version of the column is the
correct one to use.

A warning is issued if multiple matching columns are found.
}
