% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{get_seasonal_length_definition}
\alias{get_seasonal_length_definition}
\title{Get "Season Length" definition bundle}
\usage{
get_seasonal_length_definition(calculations_data, seasonal_length)
}
\arguments{
\item{calculations_data}{A list of calculation objects (R-Instat style).}

\item{seasonal_length}{Character key in \code{calculations_data}.}
}
\value{
A list with a \code{seasonal_length} element containing parsed fields
(e.g., \code{end_type}).
}
\description{
Parses the season-length definition parameters from calculation objects.
}
\seealso{
get_r_instat_definitions()
}
