% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{get_transform_column_info}
\alias{get_transform_column_info}
\title{Parse comparison expression (internal)}
\usage{
get_transform_column_info(expr)
}
\arguments{
\item{expr}{A character expression.}
}
\value{
A list with \code{direction} in \code{c("between","outer","greater","less",NA)},
and numeric \code{value}, \code{value_lb}.
}
\description{
Parses a simple comparison expression like \code{"x >= 1 & x <= 5"} and returns
\code{direction} and numeric \code{value}/\code{value_lb}.
}
\keyword{internal}
