% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{HSS}
\alias{HSS}
\title{Calculate Heidke Skill Score (HSS)}
\usage{
HSS(x, y, frcst.type, obs.type, ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Predicted values.}

\item{frcst.type}{Character. The type of forecast (e.g., "binary").}

\item{obs.type}{Character. The type of observation (e.g., "binary").}

\item{...}{Additional arguments passed to \code{verification::verify}.}
}
\value{
The Heidke skill score.
}
\description{
Computes the Heidke skill score using the \code{verification::verify} function.
}
