% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instat_comment.R
\name{instat_comment}
\alias{instat_comment}
\title{Clone \code{instat_comment} Object}
\description{
The \code{instat_comment} R6 class represents a comment in a data sheet, with various properties including identifiers, key-value pairs, comment details, timestamps, and status flags for resolution and activity.
}
\details{
Instat Comment Class

The \code{data_clone} method duplicates the current \code{instat_comment} object, ensuring any \code{instat_comment} instances within the \code{replies} field are recursively cloned. Non-\code{instat_comment} replies are directly copied without cloning.
}
\section{Methods}{

\describe{
\item{\code{data_clone(...)}}{Creates a deep clone of the current \code{instat_comment} object, including all of its fields and nested \code{instat_comment} replies.}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-instat_comment-new}{\code{instat_comment$new()}}
\item \href{#method-instat_comment-data_clone}{\code{instat_comment$data_clone()}}
\item \href{#method-instat_comment-clone}{\code{instat_comment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_comment-new"></a>}}
\if{latex}{\out{\hypertarget{method-instat_comment-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{instat_comment} object.
A comment is metadata for a row or cell of a data frame
A DataSheet will contain a list of instat_comment objects as part of the metadata for the data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_comment$new(
  id = "",
  key_values = c(),
  column = "",
  value = "",
  type = "",
  comment = "",
  label = "",
  calculation = "",
  time_stamp = "",
  replies = list(),
  resolved = FALSE,
  active = TRUE,
  attributes = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A numeric/character string representing the unique identifier for the comment. This must be unique within a data frame.}

\item{\code{key_values}}{A character vector storing key-value pairs associated with the comment. This identifies the row the comment is on.}

\item{\code{column}}{If the comment is on a cell, this is the name of the column of the cell}

\item{\code{value}}{If the comment is on a cell, this is the value in the cell at the time the comment was created.}

\item{\code{type}}{The type of comment (\code{"critical"}, \code{"warning"}, \code{"message"}, or \code{""}).}

\item{\code{comment}}{A character string for the comment text or message.}

\item{\code{label}}{A character variable. A label or grouping for the comment e.g. if comments are produced by an operation they may all have the same label. This then allows similar comments to be identified e.g. for editing/deleting}

\item{\code{calculation}}{A character variable. If the comment was created through a calculation e.g. filtering the data frame, this shows how the calculation done on the data frame}

\item{\code{time_stamp}}{The date and time (\code{POSIXct}, \code{POSIXt}) the comment was created, defaulting to the current system time if empty.}

\item{\code{replies}}{A list of replies to the comment. A reply could be a comment itself}

\item{\code{resolved}}{Logical value indicating if the comment is marked as resolved (\code{TRUE} or \code{FALSE}).}

\item{\code{active}}{Logical value indicating if the comment is marked as active (\code{TRUE} or \code{FALSE}).}

\item{\code{attributes}}{A named list of additional information about the comment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_comment-data_clone"></a>}}
\if{latex}{\out{\hypertarget{method-instat_comment-data_clone}{}}}
\subsection{Method \code{data_clone()}}{
Creates a deep clone of the current \code{instat_comment} object, including all of its fields and nested \code{instat_comment} replies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_comment$data_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{instat_comment} object with the same field values as the original, including a cloned list of \code{replies}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instat_comment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-instat_comment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{instat_comment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
