% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_climatic_element.R
\name{is_climatic_element}
\alias{is_climatic_element}
\title{Is Climatic Element}
\usage{
is_climatic_element(x)
}
\arguments{
\item{x}{Character, the name of the column.}
}
\value{
Logical, TRUE if the column is a climatic element, FALSE otherwise.
}
\description{
Check if the column name is a climatic element.
}
