% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Link Class}
\description{
The \code{link} R6 class represents a relationship between two data frames, defined by link attributes and the columns used to link them.
}
\section{Methods}{

\describe{
\item{\code{data_clone(...)}}{Creates a deep clone of the current \code{link} object, including all its fields.}
\item{\code{rename_data_frame_in_link(old_data_name, new_data_name)}}{Renames one of the data frames involved in the link.}
\item{\code{rename_column_in_link(data_name, old_column_name, new_column_name)}}{Renames a column involved in the link between data frames.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{from_data_frame}}{A character string representing the name of the first data frame in the link.}

\item{\code{to_data_frame}}{A character string representing the name of the second data frame in the link.}

\item{\code{type}}{A character string representing the type of link, e.g., "keyed".}

\item{\code{link_columns}}{A list where each element defines a way to link the data frames, with each element as a named character vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-link-new}{\code{link$new()}}
\item \href{#method-link-data_clone}{\code{link$data_clone()}}
\item \href{#method-link-rename_data_frame_in_link}{\code{link$rename_data_frame_in_link()}}
\item \href{#method-link-rename_column_in_link}{\code{link$rename_column_in_link()}}
\item \href{#method-link-clone}{\code{link$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link-new"></a>}}
\if{latex}{\out{\hypertarget{method-link-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{link} object.
Defines a relationship between two data frames and specifies linking columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{link$new(
  from_data_frame = "",
  to_data_frame = "",
  type = "",
  link_columns = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from_data_frame}}{A character string representing the name of the first data frame in the link.}

\item{\code{to_data_frame}}{A character string representing the name of the second data frame in the link.}

\item{\code{type}}{A character string representing the type of link, e.g., "keyed".}

\item{\code{link_columns}}{A list where each element defines a way to link the data frames, with each element as a named character vector. The names are columns in \code{from_data_frame} and the values are corresponding columns in \code{to_data_frame}.
Clone \code{link} Object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link-data_clone"></a>}}
\if{latex}{\out{\hypertarget{method-link-data_clone}{}}}
\subsection{Method \code{data_clone()}}{
Creates a deep clone of the current \code{link} object, including all its fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{link$data_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters to read in}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{link} object with the same field values as the original.
Rename a Data Frame in the Link.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link-rename_data_frame_in_link"></a>}}
\if{latex}{\out{\hypertarget{method-link-rename_data_frame_in_link}{}}}
\subsection{Method \code{rename_data_frame_in_link()}}{
Renames the specified data frame in the link.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{link$rename_data_frame_in_link(old_data_name, new_data_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_data_name}}{The current name of the data frame to be renamed.}

\item{\code{new_data_name}}{The new name for the data frame.
Rename a Column in the Link.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link-rename_column_in_link"></a>}}
\if{latex}{\out{\hypertarget{method-link-rename_column_in_link}{}}}
\subsection{Method \code{rename_column_in_link()}}{
Renames a column involved in the link between data frames.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{link$rename_column_in_link(data_name, old_column_name, new_column_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_name}}{The name of the data frame where the column is located.}

\item{\code{old_column_name}}{The current name of the column to be renamed.}

\item{\code{new_column_name}}{The new name for the column.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link-clone"></a>}}
\if{latex}{\out{\hypertarget{method-link-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{link$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
