% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{md}
\alias{md}
\title{Calculate Modified Index of Agreement}
\usage{
md(x, y, j = 1, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Simulated values.}

\item{j}{Numeric. Parameter for the modified index of agreement. Defaults to \code{1}.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The modified index of agreement.
}
\description{
Computes the modified index of agreement using the \code{hydroGOF::md} function.
}
