% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{PBIAS}
\alias{PBIAS}
\title{Calculate Percent Bias}
\usage{
PBIAS(x, y, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Simulated values.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The percent bias.
}
\description{
Computes the percent bias using the \code{hydroGOF::pbias} function.
}
