% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{SEDS}
\alias{SEDS}
\title{Calculate Symmetric Extreme Dependency Score}
\usage{
SEDS(x, y, frcst.type, obs.type, ...)
}
\arguments{
\item{x}{Observed values.}

\item{y}{Predicted values.}

\item{frcst.type}{Character. The type of forecast (e.g., "categorical").}

\item{obs.type}{Character. The type of observation (e.g., "categorical").}

\item{...}{Additional arguments passed to \code{verification::verify}.}
}
\value{
The symmetric extreme dependency score.
}
\description{
Computes the symmetric extreme dependency score (SEDS) using the \code{verification::verify} function.
}
