% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_country_names.R
\name{standardise_country_names}
\alias{standardise_country_names}
\title{Standardise Country Names}
\usage{
standardise_country_names(country)
}
\arguments{
\item{country}{Name of Country}
}
\value{
Name of country
}
\description{
Standardise country names in the dataset.
}
