% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_ang_dev_circular}
\alias{summary_ang_dev_circular}
\title{Calculate Angular Deviation of Circular Data}
\usage{
summary_ang_dev_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The angular deviation of the circular data.
}
\description{
Computes the angular deviation of circular data using \code{circular::angular.deviation}.
}
