% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_count}
\alias{summary_count}
\title{Count Elements in a Dataset}
\usage{
summary_count(x, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{...}{Additional arguments (not used).}
}
\value{
The count of elements in the dataset.
}
\description{
Counts the number of elements in a dataset.
}
