% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_count_missing}
\alias{summary_count_missing}
\title{Count Missing Elements in a Dataset}
\usage{
summary_count_missing(x, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{...}{Additional arguments (not used).}
}
\value{
The count of missing elements in the dataset.
}
\description{
Counts the number of missing (NA) elements in a dataset.
}
