% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_count_non_missing}
\alias{summary_count_non_missing}
\title{Count Non-Missing Elements}
\usage{
summary_count_non_missing(x, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{...}{Additional arguments (not used).}
}
\value{
The count of non-missing elements in the dataset.
}
\description{
Counts the number of non-missing (non-NA) elements in a dataset.
}
