% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_cov}
\alias{summary_cov}
\title{Calculate Covariance}
\usage{
summary_cov(
  x,
  y,
  na.rm = FALSE,
  weights = NULL,
  na_type = "",
  method = c("pearson", "kendall", "spearman"),
  use = c("everything", "all.obs", "complete.obs", "na.or.complete",
    "pairwise.complete.obs"),
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{weights}{Optional weights for the data.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{method}{Character. Covariance method ("pearson", "kendall", "spearman"). Defaults to \code{"pearson"}.}

\item{use}{Character. How missing data is handled ("everything", "all.obs", etc.). Defaults to \code{"everything"}.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The covariance or weighted covariance.
}
\description{
Computes the covariance or weighted covariance between two datasets.
}
