% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_first}
\alias{summary_first}
\title{Get First Element}
\usage{
summary_first(x, order_by = NULL, ...)
}
\arguments{
\item{x}{A vector.}

\item{order_by}{Optional vector to order by.}

\item{...}{Additional arguments (not used).}
}
\value{
The first element of the vector.
}
\description{
Returns the first element of a vector, optionally ordered by another vector.
}
