% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_kurtosis}
\alias{summary_kurtosis}
\title{Calculate Kurtosis}
\usage{
summary_kurtosis(x, na.rm = FALSE, weights = NULL, type = 2, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{weights}{Optional weights for the data.}

\item{type}{Integer. Type of kurtosis calculation. Defaults to \code{2}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The kurtosis or weighted kurtosis of the dataset.
}
\description{
Computes the kurtosis or weighted kurtosis of a dataset.
}
