% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_last}
\alias{summary_last}
\title{Get Last Element}
\usage{
summary_last(x, order_by = NULL, ...)

summary_last(x, order_by = NULL, ...)
}
\arguments{
\item{x}{A vector.}

\item{order_by}{Optional. A vector to order by before selecting the last element.}

\item{...}{Additional arguments (not used).}
}
\value{
The last element of the vector.

The last element of the vector.
}
\description{
Returns the last element of a vector, optionally ordered by another vector.

Returns the last element of a vector, optionally ordered by another vector.
}
