% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_mean_circular}
\alias{summary_mean_circular}
\title{Calculate the Mean of Circular Data}
\usage{
summary_mean_circular(
  x,
  na.rm = FALSE,
  control.circular = list(),
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{control.circular}{List of control parameters for circular objects.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The mean of the circular data.
}
\description{
Computes the mean of circular data using \code{circular::mean.circular}.
}
