% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_median_absolute_deviation}
\alias{summary_median_absolute_deviation}
\title{Calculate Median Absolute Deviation}
\usage{
summary_median_absolute_deviation(
  x,
  constant = 1.4826,
  na.rm = FALSE,
  na_type = "",
  weights = NULL,
  low = FALSE,
  high = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{constant}{Numeric. Scale factor. Defaults to \code{1.4826}.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{weights}{Optional weights for the data.}

\item{low}{Logical. Use only values below the median. Defaults to \code{FALSE}.}

\item{high}{Logical. Use only values above the median. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The median absolute deviation or weighted absolute deviation.
}
\description{
Computes the median absolute deviation or weighted absolute deviation.
}
