% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_medianHL_circular}
\alias{summary_medianHL_circular}
\title{Calculate the Hodges-Lehmann Median of Circular Data}
\usage{
summary_medianHL_circular(
  x,
  na.rm = FALSE,
  method = c("HL1", "HL2", "HL3"),
  prop = NULL,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{method}{Character string specifying the HL method ("HL1", "HL2", or "HL3").}

\item{prop}{Numeric. Proportion of data to consider.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The Hodges-Lehmann median of the circular data.
}
\description{
Computes the Hodges-Lehmann median of circular data using \code{circular::medianHL.circular}.
}
