% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_min}
\alias{summary_min}
\title{Calculate Minimum Value}
\usage{
summary_min(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The minimum value in the dataset.
}
\description{
Computes the minimum value in a dataset.
}
