% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_mode}
\alias{summary_mode}
\title{Calculate Mode}
\usage{
summary_mode(x, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{...}{Additional arguments (unused).}
}
\value{
The mode of the vector. Returns \code{NA} if the input is \code{NULL}.
}
\description{
Determines the mode (most frequent value) of a vector.
}
