% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_n_distinct}
\alias{summary_n_distinct}
\title{Count Distinct Elements}
\usage{
summary_n_distinct(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{na.rm}{Logical. Should missing values (\code{NA}) be removed? Defaults to \code{FALSE}.}

\item{...}{Additional arguments (not used).}
}
\value{
The count of distinct elements in the vector.
}
\description{
Counts the number of distinct elements in a vector.
}
