% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_nth}
\alias{summary_nth}
\title{Get nth Element}
\usage{
summary_nth(x, nth_value, order_by = NULL, ...)
}
\arguments{
\item{x}{A vector.}

\item{nth_value}{Integer. The position of the element to return.}

\item{order_by}{Optional. A vector to order by before selecting the nth element.}

\item{...}{Additional arguments (not used).}
}
\value{
The nth element of the vector.
}
\description{
Returns the nth element of a vector, optionally ordered by another vector.
}
