% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_outlier_limit}
\alias{summary_outlier_limit}
\title{Calculate Outlier Limits}
\usage{
summary_outlier_limit(
  x,
  coef = 1.5,
  bupperlimit = TRUE,
  bskewedcalc = FALSE,
  skewnessweight = 4,
  na.rm = TRUE,
  na_type = "",
  omit = FALSE,
  value = 0,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{coef}{Numeric. Coefficient for the IQR. Defaults to \code{1.5}.}

\item{bupperlimit}{Logical. Calculate upper limit if \code{TRUE}, lower limit otherwise. Defaults to \code{TRUE}.}

\item{bskewedcalc}{Logical. Use skewness in the calculation if \code{TRUE}. Defaults to \code{FALSE}.}

\item{skewnessweight}{Numeric. Weight for skewness in the calculation. Defaults to \code{4}.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{TRUE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{omit}{Logical. Omit values below a threshold. Defaults to \code{FALSE}.}

\item{value}{Numeric. Threshold for omission. Defaults to \code{0}.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The calculated outlier limit.
}
\description{
Computes the upper or lower outlier limits based on skewness and interquartile range.
}
