% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_Q3_circular}
\alias{summary_Q3_circular}
\title{Calculate the Third Quartile of Circular Data}
\usage{
summary_Q3_circular(
  x,
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{names}{Logical. Should the names of the quantiles be returned? Defaults to \code{FALSE}.}

\item{type}{Integer. Type of quantile calculation.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The third quartile of the circular data.
}
\description{
Computes the third quartile (Q3) of circular data using \code{circular::quantile.circular}.
}
