% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_quantile_circular}
\alias{summary_quantile_circular}
\title{Calculate Quantiles of Circular Data}
\usage{
summary_quantile_circular(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = FALSE,
  type = 7,
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{probs}{Numeric vector of probabilities.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{names}{Logical. Should the names of the quantiles be returned? Defaults to \code{FALSE}.}

\item{type}{Integer. Type of quantile calculation.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The quantiles of the circular data.
}
\description{
Computes the quantiles of circular data using \code{circular::quantile.circular}.
}
