% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_range}
\alias{summary_range}
\title{Calculate Range}
\usage{
summary_range(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The range of the dataset.
}
\description{
Computes the range of a dataset (difference between the maximum and minimum values).
}
