% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_range_circular}
\alias{summary_range_circular}
\title{Calculate Range of Circular Data}
\usage{
summary_range_circular(
  x,
  test = FALSE,
  na.rm = FALSE,
  finite = FALSE,
  control.circular = list(),
  na_type = "",
  ...
)
}
\arguments{
\item{x}{A vector of circular data.}

\item{test}{Logical. Perform a statistical test on the range. Defaults to \code{FALSE}.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{finite}{Logical. Should only finite values be considered? Defaults to \code{FALSE}.}

\item{control.circular}{List of control parameters for circular objects.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The range of the circular data.
}
\description{
Computes the range of circular data using \code{circular::range.circular}.
}
