% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_rho_circular}
\alias{summary_rho_circular}
\title{Calculate Rho of Circular Data}
\usage{
summary_rho_circular(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A vector of circular data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The Rho of the circular data.
}
\description{
Computes the Rho (mean resultant length) of circular data using \code{circular::rho.circular}.
}
