% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_sample}
\alias{summary_sample}
\title{Sample a Single Element}
\usage{
summary_sample(x, replace = FALSE, seed, ...)
}
\arguments{
\item{x}{A vector of data.}

\item{replace}{Logical. Should sampling be with replacement? Defaults to \code{FALSE}.}

\item{seed}{Optional. A seed for reproducibility.}

\item{...}{Additional arguments (not used).}
}
\value{
A randomly sampled element from the vector.
}
\description{
Randomly samples a single element from a vector.
}
