% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_sd}
\alias{summary_sd}
\title{Calculate Standard Deviation}
\usage{
summary_sd(x, na.rm = FALSE, weights = NULL, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{weights}{Optional weights for the data.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The standard deviation or weighted standard deviation of the data.
}
\description{
Computes the standard deviation or weighted standard deviation of a dataset.
}
