% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_skewness}
\alias{summary_skewness}
\title{Calculate Skewness}
\usage{
summary_skewness(x, weights = NULL, na.rm = FALSE, type = 2, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{Optional weights for the data.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{type}{Integer. Type of skewness calculation. Defaults to \code{2}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The skewness or weighted skewness of the dataset.
}
\description{
Computes the skewness or weighted skewness of a dataset.
}
