% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_skewness_mc}
\alias{summary_skewness_mc}
\title{Calculate Medcouple Skewness}
\usage{
summary_skewness_mc(x, na.rm = FALSE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The medcouple skewness.
}
\description{
Computes the medcouple skewness using \code{robustbase::mc}.
}
