% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_where_max}
\alias{summary_where_max}
\title{Get Corresponding Value for Maximum}
\usage{
summary_where_max(x, summary_where_y = NULL, na.rm = TRUE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{summary_where_y}{A vector of values corresponding to \code{x}.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{TRUE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
The value in \code{summary_where_y} corresponding to the maximum value in \code{x}.
}
\description{
Returns the value in another vector corresponding to the maximum value in a dataset.
}
