% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_which_min}
\alias{summary_which_min}
\title{Get Indices of Minimum Value}
\usage{
summary_which_min(x, na.rm = TRUE, na_type = "", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{TRUE}.}

\item{na_type}{Character string indicating the type of NA check to perform.}

\item{...}{Additional arguments passed to \code{na_check}.}
}
\value{
A vector of indices corresponding to the minimum value.
}
\description{
Finds all indices of the minimum value in a dataset.
}
