% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_book.R
\name{DataSheet}
\alias{DataSheet}
\title{DataSheet Class}
\description{
DataSheet Class

DataSheet Class
}
\details{
An R6 class to handle and manage a data frame with associated metadata, filters, and various settings.
}
\section{Methods}{

\describe{
\item{\code{set_data(new_data, messages, check_names)}}{Sets the data for the DataSheet object.}
\item{\code{set_changes(new_changes)}}{Sets the changes for the DataSheet object.}
\item{\code{set_filters(new_filters)}}{Sets the filters for the DataSheet object.}
\item{\code{set_column_selections(new_column_selections)}}{Sets the column selections for the DataSheet object.}
\item{\code{set_meta(new_meta)}}{Sets the metadata for the DataSheet object.}
\item{\code{clear_metadata()}}{Clears the metadata for the DataSheet object.}
\item{\code{clear_variables_metadata()}}{Clears the variables metadata for the DataSheet object.}
\item{\code{add_defaults_meta()}}{Adds default metadata to the DataSheet object.}
\item{\code{add_defaults_variables_metadata(column_names)}}{Adds default variables metadata to the DataSheet object.}
\item{\code{set_objects(new_objects)}}{Sets the objects for the DataSheet object.}
\item{\code{set_calculations(new_calculations)}}{Sets the calculations for the DataSheet object.}
\item{\code{set_keys(new_keys)}}{Sets the keys for the DataSheet object.}
\item{\code{set_comments(new_comments)}}{Sets the comments for the DataSheet object.}
\item{\code{append_to_metadata(label, value)}}{Appends to the metadata of the DataSheet object.}
\item{\code{is_metadata(label)}}{Checks if a label is in the metadata of the DataSheet object.}
}
}

\section{Active bindings}{

\describe{
\item{\code{data_changed}}{Logical, indicates if the data has changed. If setting a value, \code{new_value} must be \code{TRUE} or \code{FALSE}.}
\item{\code{metadata_changed}}{Logical, indicates if the metadata has changed. If setting a value, \code{new_value} must be \code{TRUE} or \code{FALSE}.}
\item{\code{variables_metadata_changed}}{Logical, indicates if the variables metadata has changed. If setting a value, \code{new_value} must be \code{TRUE} or \code{FALSE}.}
\item{\code{current_filter}}{A list representing the current filter. If setting a value, \code{filter} must be a list.}
\item{\code{current_column_selection}}{A list representing the current column selection. If setting a value, \code{column_selection} must be a list.}
}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data_changed}}{Logical, indicates if the data has changed.
If setting a value, new_value must be TRUE or FALSE.}

\item{\code{metadata_changed}}{Logical, indicates if the metadata has changed.
If setting a value, new_value must be TRUE or FALSE.}

\item{\code{variables_metadata_changed}}{Logical, indicates if the variables metadata has changed.
If setting a value, new_value must be TRUE or FALSE.}

\item{\code{current_filter}}{A list representing the current filter.
If setting a value, filter must be a list.}

\item{\code{current_column_selection}}{A list representing the current column selection.
If setting a value, column_selection must be a list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSheet-new}{\code{DataSheet$new()}}
\item \href{#method-DataSheet-set_data}{\code{DataSheet$set_data()}}
\item \href{#method-DataSheet-set_meta}{\code{DataSheet$set_meta()}}
\item \href{#method-DataSheet-clear_metadata}{\code{DataSheet$clear_metadata()}}
\item \href{#method-DataSheet-set_changes}{\code{DataSheet$set_changes()}}
\item \href{#method-DataSheet-set_filters}{\code{DataSheet$set_filters()}}
\item \href{#method-DataSheet-set_column_selections}{\code{DataSheet$set_column_selections()}}
\item \href{#method-DataSheet-set_objects}{\code{DataSheet$set_objects()}}
\item \href{#method-DataSheet-set_calculations}{\code{DataSheet$set_calculations()}}
\item \href{#method-DataSheet-set_keys}{\code{DataSheet$set_keys()}}
\item \href{#method-DataSheet-set_comments}{\code{DataSheet$set_comments()}}
\item \href{#method-DataSheet-clone}{\code{DataSheet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$new(
  data = data.frame(),
  data_name = "",
  variables_metadata = data.frame(),
  metadata = list(),
  imported_from = "",
  messages = TRUE,
  convert = TRUE,
  create = TRUE,
  start_point = 1,
  filters = list(),
  column_selections = list(),
  objects = list(),
  calculations = list(),
  keys = list(),
  comments = list(),
  keep_attributes = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame to be managed by the DataSheet object. Default is an empty data frame.}

\item{\code{data_name}}{A character string for the name of the data set. Default is an empty string.}

\item{\code{variables_metadata}}{A data frame containing metadata for the variables. Default is an empty data frame.}

\item{\code{metadata}}{A list containing additional metadata. Default is an empty list.}

\item{\code{imported_from}}{A character string indicating the source of the data import. Default is an empty string.}

\item{\code{messages}}{Logical, if TRUE messages will be shown during the setup. Default is TRUE.}

\item{\code{convert}}{Logical, if TRUE data will be converted. Default is TRUE.}

\item{\code{create}}{Logical, if TRUE the data will be created. Default is TRUE.}

\item{\code{start_point}}{Numeric, the starting point for default naming. Default is 1.}

\item{\code{filters}}{A list of filters to be applied to the data. Default is an empty list.}

\item{\code{column_selections}}{A list of column selections. Default is an empty list.}

\item{\code{objects}}{A list of objects associated with the data. Default is an empty list.}

\item{\code{calculations}}{A list of calculations to be performed on the data. Default is an empty list.}

\item{\code{keys}}{A list of keys for the data. Default is an empty list.}

\item{\code{comments}}{A list of comments associated with the data. Default is an empty list.}

\item{\code{keep_attributes}}{Logical, if TRUE attributes will be kept. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DataSheet} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_data}{}}}
\subsection{Method \code{set_data()}}{
Sets the data for the DataSheet object. Accepts various data types and converts them to a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_data(new_data, messages = TRUE, check_names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{The new data to be set. It can be a matrix, tibble, data.table, ts object, array, or vector.}

\item{\code{messages}}{Logical, if TRUE, messages will be shown during the data setup. Default is TRUE.}

\item{\code{check_names}}{Logical, if TRUE, column names will be checked and made valid if necessary. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The DataSheet object with the updated data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_meta"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_meta}{}}}
\subsection{Method \code{set_meta()}}{
Sets the metadata for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_meta(new_meta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_meta}}{A list containing the new metadata.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-clear_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-clear_metadata}{}}}
\subsection{Method \code{clear_metadata()}}{
Clears the metadata for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$clear_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_changes"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_changes}{}}}
\subsection{Method \code{set_changes()}}{
Sets the changes for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_changes(new_changes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_changes}}{A list containing the new changes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_filters"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_filters}{}}}
\subsection{Method \code{set_filters()}}{
Sets the filters for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_filters(new_filters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_filters}}{A list containing the new filters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_column_selections"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_column_selections}{}}}
\subsection{Method \code{set_column_selections()}}{
Sets the column selections for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_column_selections(new_column_selections)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_column_selections}}{A list containing the new column selections.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_objects"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_objects}{}}}
\subsection{Method \code{set_objects()}}{
Sets the objects for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_objects(new_objects)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_objects}}{A list containing the new objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_calculations"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_calculations}{}}}
\subsection{Method \code{set_calculations()}}{
Sets the calculations for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_calculations(new_calculations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_calculations}}{A list containing the new calculations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_keys"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_keys}{}}}
\subsection{Method \code{set_keys()}}{
Sets the keys for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_keys(new_keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_keys}}{A list containing the new keys.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
Sets the comments for the DataSheet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$set_comments(new_comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_comments}}{A list containing the new comments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSheet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSheet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSheet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
