% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{get_end_rains_definition}
\alias{get_end_rains_definition}
\title{Get "End of Rains" definition bundle}
\usage{
get_end_rains_definition(
  summary_data,
  calculations_data,
  end_rains = NULL,
  end_rains_date = NULL,
  end_rains_status = NULL,
  definitions_offset = NULL
)
}
\arguments{
\item{summary_data}{Summary data.frame used to check presence of status col.}

\item{calculations_data}{A list of calculation objects (R-Instat style).}

\item{end_rains, end_rains_date, end_rains_status}{Character names in
\code{calculations_data} to locate definitions.}

\item{definitions_offset}{The value of the offset term, found by \code{data_book$get_offset_term()}}
}
\value{
A list of parsed definition elements plus \code{output} and
\code{s_end_doy}.
}
\description{
Collects parameters defining the end-of-rains calculation and augments with
requested outputs and DOY offset.
}
\seealso{
create_end_rains_definitions(), get_r_instat_definitions()
}
