% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicsa_functions.R
\name{get_seasonal_length_definition}
\alias{get_seasonal_length_definition}
\title{Get "Season Length" definition bundle}
\usage{
get_seasonal_length_definition(calculations_data, seasonal_length)
}
\arguments{
\item{calculations_data}{A list of calculation objects (R-Instat style).}

\item{seasonal_length}{Character. Name in \code{calculations_data} for the
season-length definition.}
}
\value{
A list of parsed season-length parameters.
}
\description{
Returns parsed season-length definition parameters from calculation objects.
}
\seealso{
get_r_instat_definitions()
}
